Cenas

+++ templates/see_my_events.php
@@ -6,7 +6,7 @@
 				<li><a href="edit_event.php?id_event=<?=$row['id_event']?>">Edit event</a></li>
 				<li><a href="invite_users.php?id_event=<?=$row['id_event']?>">Invite</a></li>
 				<li><a href="">Share</a></li>
-				<li><a href="action_delete_event.php?id_event=<?=$row['id_event']?>">Delete Event</a></li>
+				<li><a href="action_delete_event.php?id_event=<?=$row['id_event']?>" onclick="confirmDelete()">Delete Event</a></li>
 			</ul>	
 			<h3>Event's name: <?= $row['name'] ?></h3>
 			<p class="type">Type: <?=$row['tipo']?></p>
@@ -19,6 +19,8 @@
 			<!-- GOD'S GIFT -->
 			<div style="clear: both"></div>
 		</div>
+		
+		<script src="confirm_delete.js"> </script>
 		
 	<?php } ?>    
   </div>
